mcp-registry/requirements.txt
numpy==1.26.3
pydantic==2.5.2
aiohttp==3.9.5
fastapi==0.100.0
uvicorn==0.23.0
pyyaml==6.0.1
python-dotenv==1.0.0
openai==1.13.3
tenacity==8.2.3
requests==2.31.0
jinja2==3.1.2 